pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "repo-distiller"
version = "0.3.0"
description = "Convert a code repository into a single text file for LLM prompting."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "Sergei Sergienko" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development",
    "Topic :: Utilities",
]

dependencies = ["typer>=0.9", "PyYAML>=6.0", "pathspec>=0.11"]

[project.scripts]
repo-distiller = "repo_distiller.cli:app"

[project.urls]
Homepage = "https://github.com/Divelix/repo-distiller"
Repository = "https://github.com/Divelix/repo-distiller"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
